replace data to one file and delete random render of cards
